Move legacy embedding utilities out of src embedding module
- keep src/utils/embedding_utils.py focused on current pipeline functions
- move legacy embedding model classes and helpers to legacy/utils/legacy_embedding_utils.py
- add legacy/utils/legacy_embedding_visualization_utils.py for plotting helpers
- update legacy visualization imports to new legacy module
